style: Refactor LoggerFSHandler implementation and file system support - Reorganized file structure and file names
